requirements.txt
Django>=4.0
djangorestframework>=3.14.0
django-cors-headers>=4.0
python-decouple>=3.0
openai>=1.0
python-dotenv>=0.19.0
anthropic>=0.3.0
Pillow>=8.0.0